Moving stripe code to it's own directory
