+++ app/javascript/src/PaymentGateways/index.js
@@ -1,2 +1,2 @@
-export { StripeFormWrapper } from 'PaymentGateways/components/StripeFormWrapper'
-export { StripeCardForm } from 'PaymentGateways/components/StripeCardForm'
+export { StripeFormWrapper } from 'PaymentGateways/stripe/components/StripeFormWrapper'
+export { StripeCardForm } from 'PaymentGateways/stripe/components/StripeCardForm'
