+++ app/javascript/src/PaymentGateways/stripe/styles/stripe.scss
